feat: Implement module seperation
